backend/requirements.txt
# Core
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.2
langchain-groq>=0.1.4
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.0

# Vector DB
chromadb>=0.5.0

# Embeddings
sentence-transformers>=2.7.0

# PDF loading
pymupdf>=1.24.0

# Pydantic
pydantic>=2.7.0
